wip(signup): style
