+++ src/features/CenteredCard.tsx
@@ -1,0 +1,13 @@
+import { Card } from "@appquality/appquality-design-system";
+import styled from "styled-components";
+
+export const CenteredCard = styled(Card)`
+  @media (min-width: ${({ theme }) => theme.grid.breakpoints.md}) {
+    max-width: 560px;
+    margin: 0 auto;
+    .card-body {
+      max-width: 300px;
+      margin: 0 auto;
+    }
+  }
+`;
