+++ src/features/LoginCard/index.tsx
@@ -1,21 +1,9 @@
-import { Card } from "@appquality/appquality-design-system";
+import { CenteredCard } from "../CenteredCard";
 import LoginForm from "../LoginForm";
-import styled from "styled-components";
-
-const StyledCard = styled(Card)`
-  @media (min-width: ${({ theme }) => theme.grid.breakpoints.md}) {
-    max-width: 560px;
-    margin: 0 auto;
-    .card-body {
-      max-width: 300px;
-      margin: 0 auto;
-    }
-  }
-`;
 
 export const LoginCard = () => {
   return (
-    <StyledCard className="login-card">
+    <CenteredCard className="login-card">
       <div className="card-body aq-text-center">
         <img
           aria-hidden
@@ -25,6 +13,6 @@ export const LoginCard = () => {
         />
         <LoginForm className="card-login-form" />
       </div>
-    </StyledCard>
+    </CenteredCard>
   );
 };
