+++ src/locales/en/translation.json
@@ -737,6 +737,7 @@
     "__RANKING_POSITION_NOLeVEL_MAX: 30": "No level"
   },
   "or log in with email and password": "",
+  "or you can": "",
   "redirecting": "redirecting",
   "social": "social",
   "testers": "Active Trybers",
