+++ src/locales/es/translation.json
@@ -737,6 +737,7 @@
     "__RANKING_POSITION_NOLeVEL_MAX: 30": "No level"
   },
   "or log in with email and password": "o inicia sesión con correo y contraseña",
+  "or you can": "",
   "redirecting": "redireccionando",
   "social": "social",
   "testers": "Tryber activos",
