+++ src/locales/it/translation.json
@@ -737,6 +737,7 @@
     "__RANKING_POSITION_NOLeVEL_MAX: 30": "No level"
   },
   "or log in with email and password": "oppure accedi con email e password",
+  "or you can": "",
   "redirecting": "reindirizzamento",
   "social": "social",
   "testers": "Tryber attivi",
